+++ src/main/java/cart/exception/PurchaseOrderException.java
@@ -1,7 +1,7 @@
 package cart.exception;
 
 import cart.domain.Member;
-import cart.domain.PurchaseOrderInfo;
+import cart.domain.purchaseorder.PurchaseOrderInfo;
 
 public class PurchaseOrderException extends RuntimeException {
     public PurchaseOrderException(String message) {
