+++ src/test/java/cart/TestFeatures.java
@@ -1,8 +1,13 @@
 package cart;
 
-import cart.domain.*;
+import cart.domain.CartItem;
+import cart.domain.Member;
+import cart.domain.Product;
 import cart.domain.point.Point;
 import cart.domain.point.UsedPoint;
+import cart.domain.purchaseorder.OrderStatus;
+import cart.domain.purchaseorder.PurchaseOrderInfo;
+import cart.domain.purchaseorder.PurchaseOrderItem;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
