+++ src/test/java/cart/application/PaymentServiceTest.java
@@ -3,10 +3,10 @@ package cart.application;
 import cart.dao.*;
 import cart.domain.CartItem;
 import cart.domain.Member;
-import cart.domain.PurchaseOrderInfo;
-import cart.domain.PurchaseOrderItem;
 import cart.domain.point.Point;
 import cart.domain.point.UsedPoint;
+import cart.domain.purchaseorder.PurchaseOrderInfo;
+import cart.domain.purchaseorder.PurchaseOrderItem;
 import cart.dto.purchaseorder.request.PurchaseOrderItemRequest;
 import cart.dto.purchaseorder.request.PurchaseOrderRequest;
 import cart.exception.PointException;
