+++ src/test/java/cart/dao/PurchaseOrderDaoTest.java
@@ -1,9 +1,9 @@
 package cart.dao;
 
 import cart.domain.Member;
-import cart.domain.OrderStatus;
 import cart.domain.Pagination;
-import cart.domain.PurchaseOrderInfo;
+import cart.domain.purchaseorder.OrderStatus;
+import cart.domain.purchaseorder.PurchaseOrderInfo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
